fix: update package.json so udacity-c3-frontend will run
The package.json that was forked from the udacity repository had older versions
of angular dependencies. When I ran the frontend, it threw an error. I found on
StackOverflow that the angular versions were not aligned with my angular cli,
so I copied/pasted all the angular dependency version numbers from a frontend
app of the same design in the udacity/cloud-developer repository.
I knew this version worked, and it has updated versions of all angular
dependencies, so I figured it would solve the problem.

After pasting the updated versions of all angular dependencies, the frontend
app works fine after 'npm install'